First draft at porting the spirit of the WorkerConf talk to written medium
I've been offline and writing stream-of-consciousness. I haven't re-read
through to check for narrative structure, consistency or even correctness, so
I've marked as draft until I can come back to it. I know for a fact that I
haven't provided enough information for twoslash to work it's magic, for example!

I have stubbed out plans for future expansion and have no ego if anyone wants
to step in and continue before I get a chance to pick up again! It's almost
certainly worth splitting this into multiple pages before going much further.

I've changed the sequencing slightly compared to the WorkerConf talk, because I
think we can do a better job of introducing concepts one-by-one, but I'm not wedded to this sequence
by any means.

Also, a question (which I might move to Discord once I'm back online): what's
the "official" name for this library? In my writing I'd like to use
capitalisation consistently to help with understanding, i.e.:

- `effect` is a common noun, as in "This function has the effect of reading from an API"
- `Effect` is the name of the type, as in "This function returns an `Effect`."

And I'm not sure how to capitalise the name of the library in order to
distinguish from both of these uses. I'd love to call it Effect-TS or something
for clarity, but that's not what it's called! An example sentence might be "Our
function reads from an API, so we represent that effect with Effect's `Effect`"